Include the other copyright holders in the LICENSE file
I am 100% sure this was just a result of auto-generating the LICENSE in
github, as you very prominently gave credit in your README (thanks!!)

I licensed my other libs as 0BSD so this is not required, but Rich Felker
licensed under MIT, so then I went ahead and did the same for all of
uwcwidth code...

Per the MIT license requirements, putting in the copyrights from all the
authors. Rich Felker in particular personally implemented the brilliantly
elegant sparse bitmap scheme and there's substantial portions of it present
in uwcwidth, so it also feels incredibly right to update this, even if it
were not a requirement :)

See https://fossa.com/blog/open-source-licenses-101-mit-license/

So, what can’t you do with MIT-licensed code? The answer is “not much,” with
a few exceptions. For one, you can’t hold the code author(s) legally liable
for any reason. You also can’t delete the copyright notice and original
license from your version of the code.

Also for example:
https://opensource.stackexchange.com/questions/13178/is-it-legal-to-remove-a-contributors-name-from-the-licence