Convert unsigned range compare into signed compares, expand tests
The originating C/Cython code uses an unsigned type and takes
advantage of a less-known trick to do a single compare for a
range:

int test_between_2_and_7(int x) {
  return (x >= 2 && x <= 7);
}

becomes:

int test_between_2_and_7(unsigned int x) {
  return (x - 2) <= (7 - 2)
}

if x == 1, the expression 1 - 2 will overflow to UINT_MAX, which
will then return 0. Indeed, 1 is outside of the range.

The advantage of this approach is that we have a single subtract
and compare vs 2 compares and a logical and.

However, in Javascript, we are not using unsigned integers and
in general these don't translate well into JS, so let's just
revert these comparisons to standard. This makes the behavior
of the JS code match the C and Cython code for some edge cases
and also hopefully makes it slightly more readable.

The added unit-tests cover these corner cases.